refactor: Make Babel plugin runtime tests Reanimated-free
