+++ apps/common-app/runtime-tests/RuntimeTestsExample.tsx
@@ -184,14 +184,12 @@ export default function RuntimeTestsExample() {
         {
           testSuiteName: 'babel plugin',
           importTest: () => {
+            require('./tests/plugin/contextObjects.test');
             require('./tests/plugin/fileWorkletization.test');
-            if (!globalThis._WORKLETS_BUNDLE_MODE_ENABLED) {
-              require('./tests/plugin/contextObjects.test');
-              require('./tests/plugin/workletClasses.test');
-            }
             require('./tests/plugin/jsxInWorklets.test');
             require('./tests/plugin/recursion.test');
             require('./tests/plugin/versionMismatch.test');
+            require('./tests/plugin/workletClasses.test');
           },
         },
         {
