--- apps/common-app/runtime-tests/tests/plugin/recursion.test.tsx
@@ -1,121 +1,0 @@
-import React, { useEffect } from 'react';
-import { View } from 'react-native';
-import { useSharedValue } from 'react-native-reanimated';
-
-import {
-  describe,
-  expect,
-  getRegisteredValue,
-  registerValue,
-  render,
-  test,
-  wait,
-} from '../../ReJest/RuntimeTestsApi';
-import { scheduleOnUI } from 'react-native-worklets';
-
-const SHARED_VALUE_REF = 'SHARED_VALUE_REF';
-
-describe('Test recursion in worklets', () => {
-  test('on Worklet runtime', async () => {
-    const ExampleComponent = () => {
-      const output = useSharedValue<number | null>(null);
-      registerValue(SHARED_VALUE_REF, output);
-      function recursiveWorklet(a: number) {
-        if (a === 2) {
-          output.value = a;
-        } else {
-          recursiveWorklet(a + 1);
-        }
-      }
-
-      useEffect(() => {
-        scheduleOnUI(recursiveWorklet, 0);
-      });
-
-      return <View />;
-    };
-    await render(<ExampleComponent />);
-    await wait(100);
-    const sharedValue = await getRegisteredValue(SHARED_VALUE_REF);
-    expect(sharedValue.onUI).toBe(2);
-  });
-
-  test('on React runtime', async () => {
-    const ExampleComponent = () => {
-      const output = useSharedValue<number | null>(null);
-      registerValue(SHARED_VALUE_REF, output);
-      function recursiveWorklet(a: number) {
-        if (a === 2) {
-          output.value = a;
-        } else {
-          recursiveWorklet(a + 1);
-        }
-      }
-
-      useEffect(() => {
-        recursiveWorklet(0);
-      });
-
-      return <View />;
-    };
-    await render(<ExampleComponent />);
-    await wait(100);
-    const sharedValue = await getRegisteredValue(SHARED_VALUE_REF);
-    expect(sharedValue.onJS).toBe(2);
-  });
-
-  test('from React to Worklet runtime', async () => {
-    const ExampleComponent = () => {
-      const output = useSharedValue<number | null>(null);
-      registerValue(SHARED_VALUE_REF, output);
-      function recursiveWorklet(a: number) {
-        if (a === 2) {
-          output.value = a;
-        } else if (a === 1) {
-          scheduleOnUI(recursiveWorklet, a + 1);
-        } else {
-          recursiveWorklet(a + 1);
-        }
-      }
-
-      useEffect(() => {
-        recursiveWorklet(0);
-      });
-
-      return <View />;
-    };
-    await render(<ExampleComponent />);
-    await wait(100);
-    const sharedValue = await getRegisteredValue(SHARED_VALUE_REF);
-    expect(sharedValue.onUI).toBe(2);
-  });
-
-  test('from Worklet to React runtime', async () => {
-    const ExampleComponent = () => {
-      const output = useSharedValue<number | null>(null);
-      registerValue(SHARED_VALUE_REF, output);
-      function recursiveWorklet(a: number) {
-        if (a === 1) {
-          output.value = a;
-        } else if (a === 2) {
-          // TODO: Such case isn't supported at the moment -
-          // a function can't be a Worklet and a Remote function at the same time.
-          // Consider supporting it in the future.
-          // scheduleOnRN(recursiveWorklet, a + 1);
-        } else {
-          recursiveWorklet(a + 1);
-        }
-      }
-
-      useEffect(() => {
-        scheduleOnUI(recursiveWorklet, 0);
-      });
-
-      return <View />;
-    };
-    await render(<ExampleComponent />);
-    await wait(100);
-    const sharedValue = await getRegisteredValue(SHARED_VALUE_REF);
-    expect(sharedValue.onJS).toBe(1);
-  });
-});
